requirements.txt
scapy
PyYAML
requests